Reformated code

+++ Student Management System/src/main/java/org/example/studentmanagementsystem/model/entities/Student.java
@@ -39,7 +39,7 @@ public class Student extends User {
         for (Grade grade : grades) {
             sum += grade.getGrade();
         }
-        return sum / grades.size();
+        return (Double) (sum / grades.size());
     }
 
 }
